+++ src/main/java/com/argprogr/portfolioweb/controller/PersonaController.java
@@ -1,7 +1,10 @@
 package com.argprogr.portfolioweb.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.argprogr.portfolioweb.dto.Mensaje;
 import com.argprogr.portfolioweb.dto.PersonaDTO;
 import com.argprogr.portfolioweb.service.PersonaService;
 
@@ -27,25 +31,36 @@ public class PersonaController {
 	public PersonaDTO getPersona(){
 		return personaService.findPersona((long) 1);
 	}
-	
+	/*
 	@PreAuthorize("hasAnyRole('ROLE_ADMIN')")
 	@PostMapping("/save")
 	public String savePersona(@RequestBody PersonaDTO dto) {
 		personaService.savePersona(dto);
 		return "Persona creada.";
 	}
-
+	*/
+	
 	@PutMapping("/edit")
-	public String updatePersona (@RequestBody PersonaDTO dto) {
+	public ResponseEntity<?> updatePersona (@RequestBody PersonaDTO dto) {
+	      if(!StringUtils.hasText(dto.getNombre())) {
+	            return new ResponseEntity(new Mensaje("El nombre es obligatorio"), HttpStatus.BAD_REQUEST);
+	        }
+	      if(!StringUtils.hasText(dto.getApellido())) {
+	            return new ResponseEntity(new Mensaje("El apellido es obligatorio"), HttpStatus.BAD_REQUEST);
+	        }
+	      
 		personaService.updatePersona((long) 1, dto);
-		return "Persona editada";
+		return new ResponseEntity("Persona editada", HttpStatus.OK);
 	}
 	
 	@PreAuthorize("hasAnyRole('ROLE_ADMIN')")
 	@DeleteMapping("/delete/{id}")
-	public String deletePersona (@PathVariable Long id) {
+	public ResponseEntity<?> deletePersona (@PathVariable Long id) {
+        if(!personaService.existsById(id)) {
+            return new ResponseEntity(new Mensaje("No existe en la base de datos."), HttpStatus.NOT_FOUND);
+            }
 		personaService.deletePersona(id);
-		return "Persona eliminada.";
+		return new ResponseEntity("Persona eliminada.", HttpStatus.OK);
 	}
 
 }
